backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
diff-match-patch==20230430
Django
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-import-export==4.1.1
django-timezone-field==7.0
django-unfold==0.39.0
django-user==0.2.1
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
djangorestframework-simplejwt==5.3.1
Faker==28.1.0
h11==0.14.0
idna==3.10
kombu==5.4.1
outcome==1.3.0.post0
packaging==24.1
pika==1.3.2
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.8
requests==2.32.3
selenium==4.25.0
sentry-sdk==2.15.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
tablib==3.5.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
whitenoise==6.7.0
wsproto==1.2.0
Django==5.1.2
